Oct 25 906am Login Issues
The channels and login look are structured correctly, except I need to
revisit contrasts on buttons.

The login is a problem. I need to combine the email and FB / Twitter
options into one succinct event.

The messages store to the database as do the users, but I need to add
the channel attribute to the database.