oct 25 905pm With Messages Channels Working for Email Login Only
Email Login moves to messages screen, channels show, they can be
selected and the messages store to the db with channel differentiated.

The FB and Twitter Auths are not moving to the messages page.  There is
a problem with the section that looks to see if the user exists already.